edit project

+++ app/Http/Controllers/AdminController.php
@@ -141,7 +141,9 @@ class AdminController extends Controller
             if ($validator->passes()) {
                 $oldImage =  DB::table('overview')->where('id', $id)->select('link_image')->first();
                 $oldImage = $oldImage->link_image;
-                File::delete('images/thum/' . $oldImage);
+                if ($oldImage != 'noimage.jpg') {
+                    File::delete('images/thum/' . $oldImage);
+                }
                 $nameImage = time() . $image->getClientOriginalName();
                 $image->move('images/thum/', $nameImage);
                 DB::table('overview')
@@ -159,6 +161,11 @@ class AdminController extends Controller
             [   'title' => $data['title'],
                 'content' => $data['content'],
                 'short_des' => $data['short_des'],
+                'square' => $data['square'],
+                'owner'  => $data['owner'],
+                'year'   => $data['year'],
+                'status' => $data['status'],
+                'address'=> $data['address']
             ]
         );
         return back()->with('response', 1);
